correcting typo in class name and reallocation testSchemaChildShouldNotHaveChildren
