+++ src/test/java/com/github/phillima/test/output/TestJSONOutput.java
@@ -1,22 +1,16 @@
 package com.github.phillima.test.output;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
 import java.io.File;
 import java.nio.file.FileSystems;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.List;
-import java.util.stream.Collectors;
 
 import com.github.phillima.asniffer.ASniffer;
 import com.github.phillima.asniffer.AmFactory;
 import com.github.phillima.asniffer.model.AMReport;
-import com.github.phillima.asniffer.model.CodeElementType;
-import com.github.phillima.asniffer.model.PackageType;
 import com.github.phillima.asniffer.output.json.d3hierarchy.Children;
 import com.github.phillima.asniffer.output.json.d3hierarchy.FetchClassViewIMP;
 import com.github.phillima.asniffer.output.json.d3hierarchy.FetchPackageViewIMP;
@@ -27,7 +21,6 @@ import com.github.phillima.asniffer.utils.ReportTypeUtils;
 
 import org.junit.BeforeClass;
 import org.junit.Ignore;
-import org.junit.Test;
 
 public class TestJSONOutput {
 
@@ -66,7 +59,6 @@ public class TestJSONOutput {
 
         //the package br.inpe.cap.output
         Children childrens = packagesContentReport.get(3).getChildByName("com.github.phillima.asniffer.output");
-
         assertEquals(1, childrens.getChildrens().size());
 
         //the package br.inpe.cap.output.json.d3hierarchy
@@ -156,33 +148,4 @@ public class TestJSONOutput {
         assertTrue(new File(dirPathResult + pathSeparator).delete());
 
     }
-
-    @Test
-    public void testSchemaChildShouldNotHaveChildren() {
-        var schemaChildTestPath = Paths.get(System.getProperty("user.dir") + "/annotationtest/schemaChildTest").toString();
-        var schemaChildReport = AmFactory.createAm(schemaChildTestPath, "asniffer").calculate();
-
-        List<Children> packagesContentReport = ReportTypeUtils.fetchPackages(schemaChildReport.getPackages(), new FetchSystemViewIMP());
-
-        var childrenOfSchema = packagesContentReport.get(0).getChildrens().stream()
-                .filter(children -> CodeElementType.SCHEMA.equals(children.getType()))
-                .filter(children -> !children.getChildrens().isEmpty())
-                .collect(Collectors.toList());
-
-        assertEquals(0, childrenOfSchema.size());
-    }
-
-    @Test
-    public void testUnnamedPackage() {
-        var unnamedPackageModel = report.getPackages().stream()
-                .filter(packageModel -> PackageType.UNNAMED.equals(packageModel.getPackageName()))
-                .findFirst()
-                .get();
-
-        var unnmedClassModel = unnamedPackageModel.getResults().get(0);
-
-        assertEquals("unnamed", unnamedPackageModel.getPackageName());
-        assertEquals("NoPackageTest", unnmedClassModel.getFullyQualifiedName());
-    }
-
 }
